remove unused func
