+++ icode/api/icode_store_api.go
@@ -23,7 +23,4 @@ import (
 type ICodeStoreApi interface {
 	//clone code from deploy info
 	Clone(id string, baseSavePath string, repositoryUrl string, sshPath string) (*icode.Meta, error)
-
-	//push code to auth repo
-	Push(meta icode.Meta) error
 }
